Modularización y creación de carpetas Products y User
