+++ src/controllers/Products/deleteProduct.js
@@ -1,4 +1,4 @@
-const { Product } = require('../db')
+const { Product } = require('../../db')
 
 async function deleteProduct(req, res) {
     const {idProduct} = req.params
