+++ src/controllers/Products/updateProduct.js
@@ -1,4 +1,4 @@
-const { Product } = require('../db')
+const { Product } = require('../../db')
 
 async function updateProduct(req, res) {
     const {id} = req.params
