+++ src/routes/index.js
@@ -3,10 +3,12 @@ const { Product } = require('../db');
 
 // Importar todos los routers;
 // Ejemplo: const authRouter = require('./auth.js');
-const { getAllProducts } = require('../controllers/getAllProducts')
-const { createProduct } = require('../controllers/createProduct')
-const { updateProduct } = require('../controllers/updateProduct')
-const { deleteProduct } = require('../controllers/deleteProduct') 
+
+
+// const { getAllProducts } = require('../controllers/getAllProducts')
+const { createProduct } = require('../controllers/Products/createProduct')
+const { updateProduct } = require('../controllers/Products/updateProduct')
+const { deleteProduct } = require('../controllers/Products/deleteProduct') 
 
 const router = Router();
 
@@ -16,7 +18,7 @@ const router = Router();
 //RUTA HACIA LANDING PAGE
 
 router.get('/', (req, res) => {
-    res.render('index')
+    res.render('Products/index')
 })
 
 //RUTA PARA VER TODOS LOS PRODUCTOS EN DB
@@ -24,7 +26,7 @@ router.get('/', (req, res) => {
 router.get('/products', async (req, res) => {
     try {
         let allProducts = await Product.findAll()
-        res.render('home', {productos:allProducts})
+        res.render('Products/home', {productos:allProducts})
     } catch (error) {
         console.log(error)
     }
@@ -33,7 +35,7 @@ router.get('/products', async (req, res) => {
 //RUTA PARA IR AL FORMULARIO DE CREACION DE PRODUCTO
 
 router.get('/crear', (req, res) => {
-    res.render('create')
+    res.render('Products/create')
 })
 
 //RUTA PARA GUARDAR EL PRODUCTO EN DB
@@ -45,7 +47,7 @@ router.post('/save', createProduct)
 router.get('/edit/:id', async (req, res) => {
     const {id} = req.params
     let productToEdit = await Product.findByPk(id)
-    res.render('edit', {product: productToEdit})
+    res.render('Products/edit', {product: productToEdit})
 })
 
 router.put('/update/:id', updateProduct)
